change communications from REST HTTP API to WebSockets
The REST HTTP API polling was creating too much conjunction and was
unreliable at time (especially in weaker networks). Hence, we changed
the whole communication set up between the hub and the device to
WebSockets. With the exception of fetching the device information at
discovery, the rest of the communication is done through a WebSocket.

We implemented a JSON protocol that tries to replicate the native
SmartThings capabilities' table structures and the offline/online logic
is now handled fully by the WebSocket connection status.

Moreover, a lot of old and no longer used functions and namings have
been modified or deleted altogether.

Signed-off-by: alon-tchelet <alon.tchelet@streamunlimited.com>